bugs in local storage

+++ client/src/pages/CartPage.tsx
@@ -1,13 +1,15 @@
-import  { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { Stack, Typography, Button } from '@mui/material';
 import ProductCartCard from '../components/ProductCartCard';
 import cartsAPI from '../api/cartsAPI';
 import { toast } from 'react-toastify';
 import { useAuth } from '../hooks/useAuth';
 import Cart from '../types/Cart';
+import * as cartLocalStorageUtils from '../utils/cartLocalStorageUtils';
+import CartItem from '../types/CartItem';
 
 const CartPage = () => {
-    const [cart, setCart] = useState<null | Cart>(null);
+    const [cartItems, setCartItems] = useState<CartItem[]>([]);
     const [loading, setLoading] = useState(true);
     const { userInfo } = useAuth();
     const [totalAmount, setTotalAmount] = useState<number>(0);
@@ -18,17 +20,14 @@ const CartPage = () => {
             try {
                 if (userInfo) {
                     const cartData = await cartsAPI.getCart();
-
-                    setCart(cartData);
+                    setCartItems(cartData.items);
                 } else {
-                    const localCart = localStorage.getItem('cart');
-                    console.log(localCart);
-                    
+                    const localCart = cartLocalStorageUtils.getCart();
+
                     if (localCart) {
-                        const localCartData = JSON.parse(localCart);
-                        setCart(localCartData);
+                        setCartItems(localCart);
                     } else {
-                        setCart(null);
+                        setCartItems([]);
                     }
                 }
             } catch (error) {
@@ -42,18 +41,26 @@ const CartPage = () => {
     }, [userInfo]);
 
     useEffect(() => {
-        if (cart) {
-            const total = cart.items.reduce((sum, item) => {
+        if (cartItems.length !== 0) {
+            const total = cartItems.reduce((sum, item) => {
                 return sum + item.quantity * item.product_id.price;
             }, 0);
             setTotalAmount(total);
         }
-    }, [cart]);
+    }, [cartItems]);
 
     const removeFromCart = async (productId: string) => {
         try {
-            const updateCart = await cartsAPI.deleteProductFromCart(productId);            
-            setCart(updateCart);
+            if (userInfo) {
+                const updateCart = await cartsAPI.deleteProductFromCart(productId);
+                
+                setCartItems(updateCart.items);
+            } else {
+                const updateCart = cartLocalStorageUtils.removeFromCart(productId);
+                const newCart = cartLocalStorageUtils.getCart();
+
+                setCartItems(newCart)
+            }
             toast.success('Product removed from cart');
         } catch (error) {
             console.error('Error removing from cart:', error);
@@ -62,41 +69,41 @@ const CartPage = () => {
     };
 
     const buyNow = async () => {
-        console.log('Product purchased!');
-        alert(`Total Amount: ${totalAmount}`);
-        const newCart = await cartsAPI.deleteCart()
-        
-        
-        setCart(newCart)
+        if (userInfo) {
+            console.log('Product purchased!');
+            alert(`Total Amount: ${totalAmount}`);
+            const newCart = await cartsAPI.deleteCart()
+            setCartItems(newCart.items)
+        } else {
+            cartLocalStorageUtils.clearCart();
+        };
+    }
+        if (loading) {
+            return <div>Loading...</div>;
+        }
+
+        if (cartItems.length === 0) {
+            return <Typography variant="h2">No items in the cart</Typography>;
+        }
+
+        return (
+            <Stack spacing={0} height={100}>
+                <Typography variant="h2" component="h2">
+                    Cart Page
+                </Typography>
+                {cartItems.map((item, index) => (
+                    <ProductCartCard
+                        key={'ProductCartCard-' + index}
+                        product={item.product_id}
+                        quantity={item.quantity}
+                        removeFromCart={removeFromCart}
+                    />
+                ))}
+                <Button variant="contained" onClick={buyNow}>
+                    Buy Now {totalAmount}
+                </Button>
+            </Stack>
+        );
     };
 
-
-    if (loading) {
-        return <div>Loading...</div>;
-    }
-
-    if (cart === null || cart.items.length === 0) {
-        return <Typography variant="h2">No items in the cart</Typography>;
-    }
-
-    return (
-        <Stack spacing={0} height={100}>
-            <Typography variant="h2" component="h2">
-                Cart Page
-            </Typography>
-            {cart.items.map((item, index) => (
-                <ProductCartCard
-                    key={'ProductCartCard-' + index}
-                    product={item.product_id}
-                    quantity={item.quantity}
-                    removeFromCart={removeFromCart}
-                />
-            ))}
-            <Button variant="contained" onClick={buyNow}>
-                Buy Now {totalAmount}
-            </Button>
-        </Stack>
-    );
-};
-
-export default CartPage;
+    export default CartPage;
